Change to float32

+++ benchmark/models.py
@@ -9,7 +9,7 @@ def isd(input_name, output_name, input_shape, output_shape, dtype):
 
 def models(batch):
     return {
-        'small_lstm': isd('inputs', 'outputs', (batch, 8, 48), (batch, 2, 12), np.float64),
+        'small_lstm': isd('inputs', 'outputs', (batch, 8, 48), (batch, 2, 12), np.float32),
         'medium_cnn': isd('inputs', 'outputs', (batch, 101, 82, 9), (batch, 101, 82, 1), np.float32),
         'large_tcnn': isd('inputs', 'outputs', (batch, 3, 101, 82, 9), (batch, 3, 101, 82, 1), np.float32)
     }
